+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Use anonymous credentials by default instead of iam metadata (use ydb.driver.credentials_from_env_variables for creds by env var)
 * Close grpc streams while closing readers/writers
 * Add control plane operations for topic api: create, drop